backend/requirements.txt
# Dependencias para el backend Django
Django>=4.2
mysqlclient>=2.2
gunicorn>=21.2
django-cors-headers>=4.4
requests>=2.28
httpx>=0.27.0
google-genai>=0.3.0
python-dotenv>=1.0.1
reportlab>=4.0.0
## NOTA: Dependencias de RAG movidas a requirements-rag.txt para reducir imagen base
